% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpkgdev-package.R
\docType{package}
\name{rpkgdev-package}
\alias{rpkgdev}
\alias{rpkgdev-package}
\title{rpkgdev: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Setor Amuzu \email{felixsetor@gmail.com} (YOUR-ORCID-ID)

}
\keyword{internal}
